[hotfix] Removes dependencies from internal roles
When consuming auto-kube-dev externally, you cause issues with dependencies
in the metadata of the internal roles. This can cause roles to be run multiple
times, which is not desired.

For now, just drop the dependency management from the metadata in the internal
roles, because it's basically assumed you're going to manage the order of
internal role firing yourself.